clasa abstracta fish cu derivate
